# app/api/admin/users/route.ts
import { NextRequest, NextResponse } from "next/server";
import { withAuth } from "@/lib/api-helpers";
import { prisma } from "@/lib/prisma";

export const GET = withAuth(["ADMIN"], async (request) => {
  const { searchParams } = new URL(request.url);
  const page = Math.max(1, Number(searchParams.get("page") ?? 1));
  const limit = Math.min(50, Math.max(1, Number(searchParams.get("limit") ?? 20)));
  const role = searchParams.get("role") ?? undefined;

  const where = role ? { role: role as "SEEKER" | "GIVER" | "ADMIN" } : {};

  const [users, total] = await Promise.all([
    prisma.user.findMany({
      where,
      orderBy: { createdAt: "desc" },
      select: {
        id: true,
        email: true,
        firstname: true,
        lastname: true,
        companyName: true,
        siret: true,
        role: true,
        banned: true,
        createdAt: true,
        _count: { select: { jobsPosted: true, applications: true } },
      },
    }),
    prisma.user.count({ where }),
  ]);

  return NextResponse.json({
    users,
    pagination: { page, limit, total, totalPages: Math.ceil(total / limit) },
  });
});