# app/api/me/route.ts
import { NextResponse } from "next/server";

import { withAuth } from "@/lib/api-helpers";
import { prisma } from "@/lib/prisma";

export const GET = withAuth(
  ["SEEKER", "GIVER"],
  async (_request, _context, session) => {
    const user = await prisma.user.findUnique({
      where: {
        id: session.user.id,
      },
      select: {
        id: true,
        firstname: true,
        lastname: true,
        email: true,
        role: true,
        companyName: true,
        siret: true,
      },
    });

    if (!user) {
      return NextResponse.json(
        { error: "Utilisateur introuvable" },
        { status: 404 }
      );
    }

    return NextResponse.json(user);
  }
);

export const PUT = withAuth(
  ["SEEKER", "GIVER"],
  async (request, _context, session) => {
    const body = await request.json();

    const {
      firstname,
      lastname,
      email,
    } = body;

    const user = await prisma.user.findUnique({
      where: {
        id: session.user.id,
      },
    });

    if (!user) {
      return NextResponse.json(
        { error: "Utilisateur introuvable" },
        { status: 404 }
      );
    }

    const updatedUser = await prisma.user.update({
      where: {
        id: session.user.id,
      },
      data: {
        firstname,
        lastname,
        email,
      },
      select: {
        id: true,
        firstname: true,
        lastname: true,
        email: true,
        role: true,
        companyName: true,
        siret: true,
      },
    });

    return NextResponse.json(updatedUser);
  }
);

export const DELETE = withAuth(
  ["SEEKER", "GIVER", "ADMIN"],
  async (_request, _context, session) => {
    const user = await prisma.user.findUnique({
      where: {
        id: session.user.id,
      },
    });

    if (!user) {
      return NextResponse.json(
        { error: "Utilisateur introuvable" },
        { status: 404 }
      );
    }

    await prisma.user.delete({
      where: {
        id: session.user.id,
      },
    });

    return new NextResponse(null, { status: 204 });
  }
);