# docs/api/openapi.geoemploi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GeoEmploi API",
    "version": "1.0.0",
    "description": "Documentation OpenAPI de l'API GeoEmploi. Les exemples ci-dessous proviennent de requêtes réellement exécutées sur l'application locale pendant le développement."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authjs.session-token"
      }
    },
...
  },
  "security": [
    { "sessionCookie": [] }
  ],
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "email": "new.user@example.com",
                "password": "MotDePasse123!",
                "firstname": "Romain",
                "lastname": "Dupont",
                "role": "SEEKER"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "example": {
                  "id": "cmtrovqot0001hhsd9en76mad",
                  "email": "new.user@example.com",
                  "role": "SEEKER"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": { "error": { "fieldErrors": { "email": ["Invalid email"] } } }
              }
            }
          },
          "409": {
            "description": "Email already exists",
            "content": {
              "application/json": {
                "example": { "error": "Cet email est déjà utilisé" }
              }
            }
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current authenticated user",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Authenticated session",
            "content": {
              "application/json": {
                "example": {
                  "id": "cmtrovq5q0000hhsdg58ur1uk",
                  "firstname": "Alice",
                  "lastname": "Giver",
                  "email": "giver.swagger@example.com",
                  "role": "GIVER"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "example": { "error": "Non authentifié" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Auth"],
        "summary": "Update current user",
        "security": [{ "sessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "firstname": "Alice",
                "lastname": "Giver",
                "email": "giver.swagger@example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "example": {
                  "id": "cmtrovq5q0000hhsdg58ur1uk",
                  "firstname": "Alice",
                  "lastname": "Giver",
                  "email": "giver.swagger@example.com",
                  "role": "GIVER"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Auth"],
        "summary": "Delete current authenticated user account",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "204": {
            "description": "Account deleted successfully"
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "example": { "error": "Non authentifié" }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "example": { "error": "Utilisateur introuvable" }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{id}": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get one job",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "Job found",
            "content": {
              "application/json": {
                "example": {
                  "id": "cmtrozih70001hhw7vx8r31np",
                  "title": "Lead Developer",
                  "description": "Mission de refonte frontend",
                  "location": "Paris 11e",
                  "latitude": 48.860162,
                  "longitude": 2.381844,
                  "salary": 60000,
                  "status": "APPROVED",
                  "giverId": "cmtrovq5q0000hhsdg58ur1uk",
                  "reportedNb": 0,
                  "createdAt": "2026-09-07T20:27:32.299Z",
                  "updatedAt": "2026-09-07T20:27:32.299Z",
                  "archived": false,
                  "giver": { "id": "cmtrovq5q0000hhsdg58ur1uk", "firstname": "Alice", "lastname": "Giver" }
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "example": { "error": "Job introuvable" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Jobs"],
        "summary": "Update a job",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "title": "Lead Developer Senior",
                "description": "Mission de refonte moderne frontend",
                "location": "Paris 11e",
                "salary": 65000
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated job",
            "content": {
              "application/json": {
                "example": {
                  "id": "cmtrozih70001hhw7vx8r31np",
                  "title": "Lead Developer Senior",
                  "description": "Mission de refonte moderne frontend",
                  "location": "Paris 11e",
                  "latitude": 48.860162,
                  "longitude": 2.381844,
                  "salary": 65000,
                  "status": "PENDING",
                  "giverId": "cmtrovq5q0000hhsdg58ur1uk",
                  "reportedNb": 0,
                  "createdAt": "2026-09-07T20:27:32.299Z",
                  "updatedAt": "2026-09-07T20:27:32.299Z",
                  "archived": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Jobs"],
        "summary": "Delete a job",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "204": { "description": "Job deleted" }
        }
      }
    },
...
    "/api/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "List users",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "role", "in": "query", "schema": { "type": "string", "enum": ["SEEKER", "GIVER", "ADMIN"] } }
        ],
        "responses": {
          "200": {
            "description": "Users list",
            "content": {
              "application/json": {
                "example": {
                  "users": [
                    {
                      "id": "cmtrovq5q0000hhsdg58ur1uk",
                      "email": "giver.swagger@example.com",
                      "firstname": "Alice",
                      "lastname": "Giver",
                      "role": "GIVER",
                      "banned": false,
                      "createdAt": "2026-09-07T20:24:36.323Z",
                      "_count": { "jobsPosted": 2, "applications": 1 }
                    }
                  ],
                  "pagination": { "page": 1, "limit": 10, "total": 9, "totalPages": 1 }
                }
              }
            }
          }
        }
      }
    },
...
  }
}